Initial AudioFileIndex stuff
