--- cpp/musicdb/audiofileindex.h
@@ -1,1 +1,0 @@
-// Stuff for the afi implemention only?
