+++ cpp/test/config_test.cpp
@@ -140,7 +140,9 @@ TEST(Config, StorageNotifcations) {
   config::clear_storage();
 }
 
+/*
 TEST(File, FilePicker) {
   auto resp = crow::response();
   files::folder_picker(resp, "Test data");
 }
+*/
